requirements.txt
fastapi
uvicorn
alembic
sqlalchemy
psycopg2-binary
python-dotenv
python-jose
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pydantic[email]
requests
pydantic_settings